Fix Changing Pray Reciter Problem In Pray Text Activity & Fix Showing Time Format

+++ app/src/main/java/com/ali/holyprays/mvp/presenter/PresenterSettingActivity.kt
@@ -36,11 +36,11 @@ class PresenterSettingActivity(
     }
 
     override fun onSavePersianFontSize(persianFontSize: Float) {
-        model.savePersianFontSize(persianFontSize.toFloat())
+        model.savePersianFontSize(persianFontSize)
     }
 
     override fun onSaveArabicFontSize(arabicFontSize: Float) {
-        model.saveArabicFontSize(arabicFontSize.toFloat())
+        model.saveArabicFontSize(arabicFontSize)
     }
 
     override fun onSaveSelectedFont(fontResId: Int) {
